AnalisisInteligenteOp/requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
google-generativeai
python-dotenv
requests
chardet